% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeMinD.R
\name{TreeMinD}
\alias{TreeMinD}
\title{Min-D test is for testing equality means against tree-ordered alternatives in one-way ANOVA}
\usage{
TreeMinD(sample_data, significance_level)
}
\arguments{
\item{sample_data}{list}

\item{significance_level}{numeric}
}
\value{
Critical value numeric

Test statistic value numeric

Result Character
}
\description{
Min-D test is for testing equality means against tree-ordered alternatives in one-way ANOVA
}
\details{
Testing of H_0:mu_0 = mu_1 = ... = mu_k vs H_1:mu_0 <= mu_i (at least one strict inequality), for all i >= 1, where mu_i represents the population means of the i-th treatment. The input consists of two variables: sample_data and significance_level. The output consists of the critical value, the Min-D test statistic value, and the result, which indicates whether to reject or not reject the null hypothesis.
}
\author{
Subha Halder
}
